Fixed review findings
Refactored SocketInspector etc.
Moved some functions and classes to separate file
